made landing page using react
