--- .gitignore
@@ -1,4 +1,0 @@
-node_modules
-.env
-package-lock.json
-
