--- README.md
@@ -1,44 +1,0 @@
-
-# List of Outrageous Excuses for Being Late to Work or School
-
-## Deployed Link :- [Weird-excuses](https://weirdest-excuses.onrender.com)
-
-## Idea
-This project allows users to share their weirdest and most outrageous excuses for being late to work or school. Users can submit their excuses, which will be listed in order from the most funny to the least funny.
-
-## Features
-- Excuse Submission: Users can type and submit their excuses for being late.
-- Excuse Rating: Other users can rate the excuses based on how funny they find them.
-- Commenting: Users can leave comments on each other's excuses, sharing their thoughts or adding to the humor.
-
-## Setting Up and Running the Project
-To set up and run the project, follow these steps:
-
-1.Clone the repository to your local machine:
-
-```
-git clone https://github.com/kalviumcommunity/S50_weird_excuses.git
-```
-
-2.Navigate to the project directory:
-```
-cd S50_weird_excuses
-```
-
-3.Install dependencies:
-```
-npm install
-```
-
-4.Start the development server:
-```
-npm run start
-```
-
-## Dependencies
-- MongoDB
-- Express
-- React
-- Node.js
-- npm
-- Vite
