--- docs.bruno/bruno.json
@@ -1,9 +1,0 @@
-{
-  "version": "1",
-  "name": "docs.bruno",
-  "type": "collection",
-  "ignore": [
-    "node_modules",
-    ".git"
-  ]
-}
