--- docs.bruno/weird-excuses.bru
@@ -1,21 +1,0 @@
-meta {
-  name: weird-excuses
-  type: http
-  seq: 1
-  description: |
-    This test checks the availability of the 'weird-excuses' API by sending a ping request to the endpoint. The 'weird-excuses' API provides a collection of hilarious excuses enetered by the users.
-}
-
-get {
-  url: https://weirdest-excuses.onrender.com/ping
-  body: none
-  auth: none
-}
-
-catch {
-  error: any
-  body: |
-    log.error("An error occurred during the HTTP request: {error.message}")
-}
-
-
