--- package.json
@@ -1,21 +1,0 @@
-{
-  "name": "weird_excuses",
-  "version": "1.0.0",
-  "description": "A collection of weird excuses for being late for work or school.",
-  "main": "index.js",
-  "scripts": {
-    "start": "nodemon server.js",
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "author": "prathamshailesh",
-  "license": "ISC",
-  "dependencies": {
-    "dotenv": "^16.4.1",
-    "express": "^4.18.2",
-    "express-validator": "^7.0.1",
-    "mongoose": "^8.1.2"
-  },
-  "devDependencies": {
-    "nodemon": "^3.0.3"
-  }
-}
